requirements.txt
django==4.2.1
mutagen==1.46.0
deepl
python-dotenv
openai
pydub
djangorestframework
django-rest-swagger
drf-yasg
django-cors-headers
moviepy
google-cloud-storage
django-storages
dj-rest-auth
dj-rest-auth[with_social]